Improve ResourceReader's API, add better docs
