+++ src/lib.rs
@@ -19,6 +19,7 @@ mod template;
 mod tile;
 mod tileset;
 mod util;
+mod reader;
 
 pub use animation::*;
 pub use cache::*;
@@ -32,3 +33,4 @@ pub use properties::*;
 pub use template::*;
 pub use tile::*;
 pub use tileset::*;
+pub use reader::*;
